requirements.txt
torch
torchvision
transformers
pillow
easyocr
faiss-cpu
numpy
